move js utils
